The API now takes online request according to the format specified in the Servlet class
Next order of business is to include the decrementing parameter as described in todo: issue 1 (also in the online servlet)

Online algorithm: the new point will be placed in the index of the closest point to in the MAIN POINTS array, pushing the existing point 1 place further. Hazard: source point may be pushed to be second. so maybe do that the other way around? pushing the new point next
Details in the todo